readnwin-backend/requirements.txt
aiofiles==24.1.0
alembic==1.12.1
annotated-types==0.7.0
anyio==3.7.1
backports-datetime-fromisoformat==2.0.3
bcrypt==4.0.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.8
cryptography==45.0.6
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
environs==14.3.0
exceptiongroup==1.3.0
fastapi==0.104.1
fastapi_cors==0.0.6
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==4.0.0
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
python-decouple==3.8
python-dotenv==1.1.1
python-jose==3.3.0
python-multipart==0.0.6
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.24.0
reportlab==4.0.7
Pillow==10.0.1
